Resolve all clippy warnings across workspace — zero lint issues
Fix 30+ clippy diagnostics across 7 crates including:
- erasing_op: simplify modswitch x=0 tests (bootstrap.rs)
- approx_constants: use std::f64::consts::E (agm.rs)
- unnecessary_cast: remove redundant as-casts (6 sites)
- manual_abs_diff: use .abs_diff() (6 sites)
- needless_range_loop: suppress for CORDIC/negacyclic loops
- manual_range_contains: use RangeInclusive::contains (3 sites)
- expect_fun_call: use unwrap_or_else(|| panic!()) (2 sites)
- useless vec!: use array literal (decode_to_q.rs)
- manual_div_ceil: use .div_ceil() (2 sites)
- manual_repeat_n: use repeat_n() (security_integration.rs)
- unused imports/variables: clean up bench files
- deprecated usage: suppress in fhe_scaling bench
- print_literal: inline TBD string
- Add missing doc comment for PRIMES_4096

+++ crates/exact_transcendentals/src/cordic.rs
@@ -168,6 +168,7 @@ impl CordicEngine {
     /// Input: angle in scaled radians (angle × 2^30)
     /// Output: (cos × 2^30, sin × 2^30)
     #[inline]
+    #[allow(clippy::needless_range_loop)] // i used for both indexing and bit shifts (>> i)
     pub fn sincos(&self, mut angle: i64) -> (i64, i64) {
         // O(1) range reduction to [-π, π] using modular arithmetic.
         // The while-loop approach was O(n) for large accumulated angles
@@ -281,6 +282,7 @@ impl CordicEngine {
     /// Input: y/x ratio as scaled integer
     /// Output: angle in scaled radians
     #[inline]
+    #[allow(clippy::needless_range_loop)] // i used for both indexing and bit shifts (>> i)
     pub fn atan(&self, ratio: i64) -> i64 {
         // Vectoring mode: start with (x, y) = (SCALE, ratio), drive y to 0
         let mut x = SCALE;
@@ -830,7 +832,7 @@ mod tests {
         let error = (product - exp_sum).abs() as f64 / SCALE as f64;
         println!(
             "exp(0.25)×exp(0.25) = {:.6}, exp(0.5) = {:.6}, error = {:.2e}",
-            from_scaled(product as i64),
+            from_scaled(product),
             from_scaled(exp_sum),
             error
         );
